requirements.txt
Appium-Python-Client==4.5.1
selenium==4.29.0
pytest==8.3.4